@@ -2,6 +2,7 @@
   "name": "form-delegate",
   "version": "0.0.1",
   "repository": {},
+  "license": "GPL-3.0",
   "dependencies": {
     "babel-polyfill": "^6.16.0",
     "bourbon": "beta",
@@ -17,9 +18,9 @@
     "phoenix": "file:deps/phoenix",
     "phoenix_html": "file:deps/phoenix_html",
     "query-string": "^5.0.0",
-    "react": "^15.6.1",
+    "react": "^16.0.0",
     "react-copy-to-clipboard": "^5.0.0",
-    "react-dom": "^15.1.0",
+    "react-dom": "^16.0.0",
     "react-helmet": "^5.1.3",
     "react-redux": "^5.0.1",
     "react-router-dom": "^4.1.1",
@@ -33,11 +34,11 @@
   "devDependencies": {
     "autoprefixer": "^7.1.1",
     "babel-core": "^6.9.1",
-    "babel-eslint": "^7.2.0",
-    "babel-jest": "^20.0.3",
+    "babel-eslint": "^8.0.1",
+    "babel-jest": "^21.2.0",
     "babel-loader": "^7.1.0",
     "babel-polyfill": "^6.16.0",
-    "babel-preset-es2015": "^6.16.0",
+    "babel-preset-env": "^1.6.1",
     "babel-preset-react": "^6.16.0",
     "babel-preset-stage-0": "^6.5.0",
     "copy-webpack-plugin": "^4.0.0",
@@ -50,17 +51,17 @@
     "eslint-plugin-prettier": "^2.1.2",
     "eslint-plugin-react": "^7.1.0",
     "extract-text-webpack-plugin": "^3.0.0",
-    "file-loader": "^0.11.2",
+    "file-loader": "^1.1.5",
     "husky": "^0.14.1",
-    "jest": "^20.0.4",
+    "jest": "^21.2.1",
     "lint-staged": "^4.0.0",
     "node-sass": "^4.5.3",
     "postcss-loader": "^2.0.6",
     "prettier": "^1.5.2",
-    "react-test-renderer": "^15.3.2",
+    "react-test-renderer": "^16.0.0",
     "sass-loader": "^6.0.0",
-    "style-loader": "^0.18.2",
-    "url-loader": "^0.5.7",
+    "style-loader": "^0.19.0",
+    "url-loader": "^0.6.2",
     "webpack": "^3.4.1"
   },
   "lint-staged": {
