@@ -1,4 +1,5 @@
-import React, { PropTypes } from 'react';
+import React from 'react';
+import PropTypes from 'prop-types';
 import { Field } from 'redux-form';
 import { last } from 'lodash';
 import FormIntegrationList from '../components/FormIntegrationList';
