@@ -1,4 +1,5 @@
-import React, { PropTypes } from 'react';
+import React from 'react';
+import PropTypes from 'prop-types';
 import { map } from 'lodash';
 import Message from '../components/Message';
 import moment from 'moment';
