@@ -1,4 +1,5 @@
-import React, { PropTypes } from 'react';
+import React from 'react';
+import PropTypes from 'prop-types';
 import { animateScroll } from 'react-scroll';
 import { browserHistory } from 'react-router-dom';
 import { connect } from 'react-redux';
