@@ -19,7 +19,7 @@ module.exports = {
         exclude: /node_modules/,
         loader: 'babel-loader',
         options: {
-          presets: ['es2015', 'stage-0', 'react'],
+          presets: ['env', 'stage-0', 'react'],
         },
       },
       {
